Global self-assignment of Id [by TR3E]
If two or more components match the same identification, the application will die horribly. If it is not at the beginning of the activity, it will be when you restart the application and restore your activity. In this way, LAMW manages the ids, preventing any component from having the same id.

I have assigned a range from 1 to 100000 for component ids assigned from Lazarus and 100000 onwards for components assigned by Java. It can be changed if necessary in "Controls.java".

I have tested this change in different LAMW sample applications by adapting the code of JDBListView, jExpanedableListView, jListView, jModalDialog, jNumberPickerDialog, jRadioGroup, jsNavigationView, jsRecyclerView for proper operation and assignment of the correct ids.