Revised and updated jImageView, jImageBtn, jImageList [by TR3E]
Revised and updated jImageView, jImageBtn, jImageList [by TR3E]

Added event to AndroidForm "OnInit" to execute code when conducting Init.

Added event on jImageBtn "OnDown"

Updated the code so that when loading images from assets you can define the density of these images and thus display correctly in different screen densities. To do this, it is the function "gapp.SetDensity (daHIGH)", for example, which must be called in the "OnInit" event to define the density of the images to be loaded before being loaded.