Added support for recreating activities [by TR3E]
Added support for recreating activities [by TR3E]

Added the "OnAppActivityRecreate" method that we can call to initialize variables or other things when detecting activity recreation.

Android reserves the right to recreate activities at any time, LAMW when this happens was blocked, with this update resolved.

For example, in Android 8 and 7 in several windows when resizing, sometimes the activity is recreated.