Updated "jsAdmob" with new announcements [by ADiV]
Updated "jsAdmob" with new announcements.

Now it has support for Rewarded and Interstacional announcements, it is updated to version 19.3.0

Notes for updating a program that already uses old AdMob:
You need to update some functions that have been adapted to accommodate the new types of ads and events have to be recreated by removing the old ones and generating new ones.