Update to support Gradle 4.1.0 [by ADiV]
With this update we will be able to use the minimum Gradle 4.1.0 plugin recommended by Google and we will be able to activate the native debugging code.

In an already created app we need to remove android:minSdkVersion="16" from "AndroidManifest.xml" which is not necessary in gradle since we use build.gradle. If it is not deleted, it generates a compilation error. Then "Clean and Build" and "Build Android Apk"

Update to support Gradle 4.1.0:
classpath 'com.android.tools.build:gradle:4.1.0'
ndk { debugSymbolLevel 'FULL' }